+++ pages/_app.tsx
@@ -1,13 +1,11 @@
 import Footer from '@/components/Footer/Footer';
 import Nav from '@/components/Nav/Nav';
 import { ModalProvider } from '@/contexts/ModalContext';
-import { UserContext } from '@/contexts/UserContext';
+import { UserProvider } from '@/contexts/UserContext';
 import '@/styles/globals.css';
 import type { AppProps } from 'next/app';
 import Head from 'next/head';
-import { useEffect, useState } from 'react';
-import { getUser } from '../api/api';
-import { User } from '@/hooks/useGetUser';
+import { useEffect } from 'react';
 
 declare global {
   interface Window {
@@ -16,26 +14,6 @@ declare global {
 }
 
 export default function App({ Component, pageProps }: AppProps) {
-  const [user, setUser] = useState<User>({
-    id: '',
-    created_at: new Date(),
-    name: '',
-    image_source: '',
-    email: '',
-    auth_id: '',
-  });
-
-  useEffect(() => {
-    const userAccess = localStorage.getItem('token');
-    if (userAccess) {
-      const loadUser = async () => {
-        const response = await getUser(userAccess);
-        setUser(response[0]);
-      };
-      loadUser();
-    }
-  }, []);
-
   useEffect(() => {
     window.Kakao.init(process.env.NEXT_PUBLIC_KAKAO_KEY);
   }, []);
@@ -45,12 +23,12 @@ export default function App({ Component, pageProps }: AppProps) {
       <Head>
         <title>Linkbrary</title>
       </Head>
-      <UserContext.Provider value={user.id}>
-        <Nav user={user} />
+      <UserProvider>
+        <Nav />
         <ModalProvider>
           <Component {...pageProps} />
         </ModalProvider>
-      </UserContext.Provider>
+      </UserProvider>
       <Footer />
     </>
   );
